+++ lib/src/apollovm_generator.dart
@@ -129,6 +129,8 @@ abstract class ApolloGenerator<
       return generateASTBranch(statement, out: out);
     } else if (statement is ASTStatementForLoop) {
       return generateASTStatementForLoop(statement, out: out);
+    } else if (statement is ASTStatementForEach) {
+      return generateASTStatementForEach(statement, out: out);
     } else if (statement is ASTStatementReturnNull) {
       return generateASTStatementReturnNull(statement, out: out);
     } else if (statement is ASTStatementReturnValue) {
@@ -157,6 +159,8 @@ abstract class ApolloGenerator<
   }
 
   O generateASTStatementForLoop(ASTStatementForLoop forLoop, {O? out});
+
+  O generateASTStatementForEach(ASTStatementForEach forEach, {O? out});
 
   O generateASTBranchIfBlock(ASTBranchIfBlock branch, {O? out});